requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
amqp==5.2.0; python_version >= '3.6'
annotated-types==0.7.0; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
anyio==4.4.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
billiard==4.2.0; python_version >= '3.7'
celery==5.4.0; python_version >= '3.8'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dirtyjson==1.0.8
distro==1.9.0; python_version >= '3.6'
django==5.0.6; python_version >= '3.10'
django-filter==24.2; python_version >= '3.8'
django-restframework==0.0.1
djangorestframework==3.15.2; python_version >= '3.8'
drf-spectacular==0.27.2; python_version >= '3.7'
filelock==3.15.4; python_version >= '3.8'
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.6.1; python_version >= '3.8'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==22.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
importlib-metadata==8.0.0; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
joblib==1.4.2; python_version >= '3.8'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu==5.3.7; python_version >= '3.8'
langchain-core==0.2.11; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.1.14; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.1.84; python_version < '4.0' and python_full_version >= '3.8.1'
llama-cloud==0.0.6; python_version >= '3.8' and python_version < '4'
llama-index==0.10.52; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-agent-openai==0.2.7; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-cli==0.1.12; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-core==0.10.52.post1; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-embeddings-openai==0.1.10; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-indices-managed-llama-cloud==0.2.3; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-legacy==0.9.48; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-llms-openai==0.1.25; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-multi-modal-llms-openai==0.1.7; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-program-openai==0.1.6; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-question-gen-openai==0.1.3; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-readers-file==0.1.28; python_version < '4.0' and python_full_version >= '3.8.1'
llama-index-readers-llama-parse==0.1.6; python_version < '4.0' and python_full_version >= '3.8.1'
llama-parse==0.4.6; python_version < '4.0' and python_full_version >= '3.8.1'
marshmallow==3.21.3; python_version >= '3.8'
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.3; python_version >= '3.10'
nltk==3.8.1; python_version >= '3.7'
numpy==1.26.4; python_version >= '3.9'
openai==1.35.10; python_full_version >= '3.7.1'
orjson==3.10.6; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pathvalidate==3.2.0; python_version >= '3.7'
phonenumbers==8.13.40
pillow==10.4.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycountry==24.6.1; python_version >= '3.8'
pydantic==2.8.2; python_full_version >= '3.12.4'
pydantic-core==2.20.1; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pypdf==4.2.0; python_version >= '3.6'
python-arango==8.0.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.7; python_version >= '3.7'
referencing==0.35.1; python_version >= '3.8'
regex==2024.5.15; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-file==2.1.0
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.19.0; python_version >= '3.8'
rstr==3.2.2; python_version >= '3.7'
schwifty==2024.6.1; python_version >= '3.8'
setuptools==70.2.0; python_version >= '3.8'
simplejson==3.19.2; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy[asyncio]==2.0.31; python_version >= '3.7'
sqlparse==0.5.0; python_version >= '3.8'
stix2==3.0.1; python_version >= '3.6'
stix2-patterns==2.0.0; python_version >= '3.6'
striprtf==0.0.26
tenacity==8.5.0; python_version >= '3.8'
tiktoken==0.7.0; python_version >= '3.8'
tld==0.13; python_version >= '3.7' and python_version < '4'
tldextract==5.1.2; python_version >= '3.8'
tqdm==4.66.4; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2024.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
validators==0.31.0; python_version >= '3.8'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
zipp==3.19.2; python_version >= '3.8'

https://github.com/muchdogesec/txt2stix/archive/obstracts.zip
https://github.com/muchdogesec/file2txt/archive/beta.zip
https://github.com/muchdogesec/stix2arango/archive/main.zip